requirements.txt
llama-index>=0.10.0
llama-index-embeddings-huggingface>=0.2.0
llama-index-vector-stores-chroma>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
torch>=2.0.0
